Revert "164"
This reverts commit 257d841f7d795cc4c8c63037d333ac5fc48818c9.

+++ drivers/cdrom/cdrom.c
@@ -2425,7 +2425,7 @@ static int cdrom_ioctl_select_disc(struct cdrom_device_info *cdi,
 		return -ENOSYS;
 
 	if (arg != CDSL_CURRENT && arg != CDSL_NONE) {
-		if (arg >= cdi->capacity)
+		if ((int)arg >= cdi->capacity)
 			return -EINVAL;
 	}
